refactor: extract CommandOutputStyle and ParsesMediaIds console traits
- Add `src/Console/Concerns/CommandOutputStyle.php` with `section()`, `statusLine()`, `formatBytes()` and the `$hasErrors` tracking property — matches Laravel's `about` command layout
- Add `src/Console/Concerns/ParsesMediaIds.php` with `parseMediaIds()` supporting comma-separated IDs and ranges (`1,3,5..10`)
- Apply both traits to `DoctorCommand` and `GenerateConversionsCommand`, removing the duplicated inline helpers
- No behavior change; output and CLI signatures unchanged
